move new intrinsics from i586 to i686 module
